# docs/swagger.yaml
basePath: /
...
host: localhost:8080
info:
  contact:
    email: vladislavtagaev@gmail.com
    name: ABC Helper Team
    url: https://github.com/Sales-Analysis/abc-helper-backend
  description: Minimal HTTP API for abc-helper-backend (hello, probes, version).
  title: ABC Helper Backend API
  version: 0.1.0
paths:
...
  /api/v1/abc/upload:
    post:
      consumes:
      - multipart/form-data
      description: Accepts an XLSX file, validates it and returns OK if valid.
      parameters:
      - description: XLSX file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Upload ABC XLSX
      tags:
      - analysis
...
swagger: "2.0"